Move Pods creation logic to conftest
