+++ tests/integration/conftest.py
@@ -1,13 +1,16 @@
-import logging
-from pathlib import Path
-
 import pytest
+import logging
+
+from pathlib import Path
 from lightkube import Client, codecs, KubeConfig
+from lightkube.resources.apps_v1 import DaemonSet
+from lightkube.resources.core_v1 import Pod
+
 
 log = logging.getLogger(__name__)
 
 
-@pytest.fixture()
+@pytest.fixture(scope="module")
 async def kubeconfig(ops_test):
     kubeconfig_path = ops_test.tmp_path / "kubeconfig"
     rc, stdout, stderr = await ops_test.run(
@@ -25,7 +28,7 @@ async def kubeconfig(ops_test):
     yield kubeconfig_path
 
 
-@pytest.fixture()
+@pytest.fixture(scope="module")
 async def client(kubeconfig):
     config = KubeConfig.from_file(kubeconfig)
     client = Client(
@@ -36,11 +39,30 @@ async def client(kubeconfig):
 
 
 @pytest.fixture()
-def iperf3_yaml_path(client):
+def iperf3_pods(client):
     path = Path.cwd() / "tests/data/iperf3_daemonset.yaml"
-    yield path
+    with open(path) as f:
+        for obj in codecs.load_all_yaml(f):
+            client.create(obj)
+
+    wait_daemonset(client, "ls1", "perf", 3)
+    pods = list(client.list(Pod, namespace="ls1"))
+
+    yield pods
+
     with open(path) as f:
         for obj in codecs.load_all_yaml(f):
             client.delete(
                 type(obj), obj.metadata.name, namespace=obj.metadata.namespace
             )
+
+
+def wait_daemonset(client: Client, namespace, name, pods_ready):
+    for _, obj in client.watch(
+        DaemonSet, namespace=namespace, fields={"metadata.name": name}
+    ):
+        if obj.status is None:
+            continue
+        status = obj.status.to_dict()
+        if status["numberReady"] == pods_ready:
+            return
